Add context and in-between rw accessors
Sometimes we duplicate rws:

	(f    (x y) (y))

is a common pattern, so now we can write

	(f    (x) y)

This saves us the verbosity.

In addition, we now have context that is added to the reading part of the equation, for instance:

	(f    (x a b c))
	(g    (x d e f))
	(h    (x g h i))

Is a common pattern, now we can use

	(context x
	  (f    (a b c))
	  (g    (d e f))
	  (h    (g h i)))

Note that these features are undocumented and experimental, they may be altered later.