Add controllers
Controllers are derived concepts built on top of spipe. These are
currently experimental. Instead of creating a new function that loops,
we can instead use a simple construct that loops through a list of
lambdas:

        (H~>
                state
                ((loop~> some access))
                )

This allows the code to be clearer. There is nothing magical about this
at all. `loop~>` is simply a macro that generates a lambda that operates
on the entire state.

In the future, we may add these as keywords to make the prettier to look
at:

        (H~>
                state
                (#:loop some.list.of.lambdas)
                )