Restart the loop~> upon a change
This way of operating the loop~> is most likely the best and simplest
way. What we do now is that - if the invocation has changed the list in
any way, we just restart the loop completely. If it is unchanged, only
then do we iterate the loop.

This ensures that a pop will correctly call the function after the
current one, and a push will call the new function immediately, however,
it will re-run the function that pushed it. But that's probably fine...