Add opinionated ideas about an fsm
This is a transitional update, I really want to test this out. It's an
ad-hoc addition of an fsm-like system to spipe, which is useful for
dynamically running other spipes.

Currently, we store stuff in a top-level variable called 'fsm. Not the
best name, but it works... Here's an example

        (define/H~> transformation
          ((const (list trans2))  some-list)
          ((const 1)              value)
          ((top~> some-list)))

        (define/H~> trans2
          (displayln (value)))

`top~>` simply runs the top element of some-list.